+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/XR/XRPass.cs
@@ -182,7 +182,7 @@ namespace UnityEngine.Rendering.HighDefinition
                 {
                     Debug.LogWarning("If you're trying to enable XR single-pass after the first frame, you need to set TextureXR.maxViews to 2 before the render pipeline is created (typically in a script with Awake()).");
                 }
-                
+
                 throw new NotImplementedException($"Invalid XR setup for single-pass, trying to add too many views! Max supported: {maxSupportedViews}");
             }
         }
@@ -255,7 +255,7 @@ namespace UnityEngine.Rendering.HighDefinition
             }
         }
 
-        internal void RenderOcclusionMeshes(CommandBuffer cmd, RTHandle depthBuffer)
+        internal void RenderOcclusionMeshes(CommandBuffer cmd, Color clearColor, RTHandle colorBuffer, RTHandle depthBuffer)
         {
             if (enabled && xrSdkEnabled && occlusionMeshMaterial != null)
             {
@@ -267,11 +267,37 @@ namespace UnityEngine.Rendering.HighDefinition
                     {
                         if (views[viewId].occlusionMesh != null)
                         {
-                            CoreUtils.SetRenderTarget(cmd, depthBuffer, ClearFlag.None, 0, CubemapFace.Unknown, viewId);
+                            CoreUtils.SetRenderTarget(cmd, colorBuffer, depthBuffer, ClearFlag.None, clearColor, 0, CubemapFace.Unknown, viewId);
+                            cmd.SetGlobalVector(HDShaderIDs._ClearColor, clearColor);
                             cmd.DrawMesh(views[viewId].occlusionMesh, m, occlusionMeshMaterial);
                         }
                     }
                 }
+            }
+        }
+
+        static readonly int _unity_StereoMatrixV = Shader.PropertyToID("unity_StereoMatrixV");
+        static readonly int _unity_StereoMatrixP = Shader.PropertyToID("unity_StereoMatrixP");
+        static readonly int _unity_StereoMatrixVP = Shader.PropertyToID("unity_StereoMatrixVP");
+        Matrix4x4[] builtinViewMatrix = new Matrix4x4[2];
+        Matrix4x4[] builtinProjMatrix = new Matrix4x4[2];
+        Matrix4x4[] builtinViewProjMatrix = new Matrix4x4[2];
+
+        // Maintain compatibility with builtin renderer
+        internal void UpdateBuiltinStereoMatrices(CommandBuffer cmd)
+        {
+            if (singlePassEnabled)
+            {
+                for (int viewIndex = 0; viewIndex < 2; ++viewIndex)
+                {
+                    builtinViewMatrix[viewIndex] = GetViewMatrix(viewIndex);
+                    builtinProjMatrix[viewIndex] = GL.GetGPUProjectionMatrix(GetProjMatrix(viewIndex), true);
+                    builtinViewProjMatrix[viewIndex] = builtinProjMatrix[viewIndex] * builtinViewMatrix[viewIndex];
+                }
+
+                cmd.SetGlobalMatrixArray(_unity_StereoMatrixV, builtinViewMatrix);
+                cmd.SetGlobalMatrixArray(_unity_StereoMatrixP, builtinProjMatrix);
+                cmd.SetGlobalMatrixArray(_unity_StereoMatrixVP, builtinViewProjMatrix);
             }
         }
     }