+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New built-in operator providing new times access
 - More efficient update modes inspector
 - Ability to read attribute in spawn context through graph
+- Added save button to save only the current visual effect graph.
+- Added Degrees / Radians conversion subgraphs in samples
 
 ### Fixed
 - Moved VFX Event Tester Window visibility to Component Play Controls SceneView Window